+++ index.html
@@ -8,7 +8,7 @@
 		<meta
 			name="viewport"
 			content="width=device-width, initial-scale=1.0" />
-		<title>Vite App</title>
+		<title>科协聊天室</title>
 		<style>
 			* {
 				margin: 0;
